+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Upgrade the `com.google.testing.compile` artifact to version `0.18-rf`.
 * Upgrade the `org.realityforge.proton` artifacts to version `0.32`.
 * Upgrade the `org.realityforge.guiceyloops` artifact to version `0.106`.
+* Allow wildcard type parameter on the `ComputableValue` instance returned from a `@ComputableValueRef` annotated method.
 
 ### [v0.166](https://github.com/arez/arez/tree/v0.166) (2020-01-16) · [Full Changelog](https://github.com/arez/arez/compare/v0.165...v0.166)